Whisper_Python/requirements.txt
# Whisper.cpp Web App - Dependências

# Framework Web
flask==2.3.3
werkzeug==2.3.7

# Processamento de Áudio
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1
librosa==0.10.1

# Processamento de Vídeo/Áudio
ffmpeg-python==0.2.0

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Interface com Whisper.cpp
subprocess32==3.5.4; python_version < '3.0'

# Segurança
pyopenssl==23.2.0